requirements.txt
amqp==5.1.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
kombu==5.3.2
mysql-connector-python==8.1.0
outcome==1.2.0
prompt-toolkit==3.0.39
protobuf==4.21.12
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.0
schedule==1.2.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
